+++ apps/judicial-system/web/src/routes/Login/Login.tsx
@@ -1,6 +1,11 @@
 import React from 'react'
 
-import { Typography, Button, Box, Alert } from '@island.is/island-ui/core'
+import {
+  Typography,
+  Button,
+  Box,
+  AlertMessage,
+} from '@island.is/island-ui/core'
 import { apiUrl } from '../../api'
 import * as styles from './Login.treat'
 
@@ -12,7 +17,7 @@ export const Login = () => {
       {urlParams.has('error') && (
         <div className={styles.errorMessage}>
           <Box marginBottom={6}>
-            <Alert
+            <AlertMessage
               type="info"
               title="Innskráning ógild"
               message="Innskráning ekki lengur gild. Vinsamlegast reynið aftur."
