+++ libs/island-ui/core/src/index.ts
@@ -34,14 +34,13 @@ export * from './lib/Link'
 export * from './lib/Tooltip/Tooltip'
 export * from './lib/InputError/InputError'
 export * from './lib/FieldPolarQuestion/FieldPolarQuestion'
-export * from './lib/Badge/Badge'
 export * from './lib/Divider/Divider'
 export * from './lib/LinkCard/LinkCard'
 export * from './lib/Tag/Tag'
 export * from './lib/AsyncSelect/AsyncSelect'
 export * from './lib/SkeletonLoader/SkeletonLoader'
 export * from './lib/Blockquote/Blockquote'
-export * from './lib/Alert/Alert'
+export * from './lib/AlertMessage/AlertMessage'
 export { Swiper } from './lib/Swiper'
 export { FocusableBox } from './lib/FocusableBox'
 export { NewsletterSignup } from './lib/NewsletterSignup'
